meshgraphnets-pytorch/requirements.txt
h5py
matplotlib
numpy
opencv
Pillow
pytorch
pytorch_geometric
pytorch-scatter
tqdm